components: rename stats cards
